completed contact manager api

+++ controllers/contactController.js
@@ -3,15 +3,15 @@ const Contact = require("../model/contactModel")
 
 // @desc get contacts
 // @route GET /api/contacts
-// access public
+// access private
 const getContacts = asyncHandler(async (req, res) => {
-  const contacts = await Contact.find()
+  const contacts = await Contact.find({ user_id: req.user.id })
   res.status(200).json(contacts)
 })
 
 // @desc create contact
 // @route POST /api/contacts
-// access public
+// access private
 const createContact = asyncHandler(async (req, res) => {
   console.log(`contact body is: ${req.body}`)
   const { name, email, phone } = req.body
@@ -23,13 +23,14 @@ const createContact = asyncHandler(async (req, res) => {
     name,
     email,
     phone,
+    user_id: req.user.id,
   })
   res.status(201).json(contact)
 })
 
 // @desc get contact
 // @route GET /api/contacts/:id
-// access public
+// access private
 const getContact = asyncHandler(async (req, res) => {
   const contact = await Contact.findById(req.params.id)
   if (!contact) {
@@ -41,29 +42,37 @@ const getContact = asyncHandler(async (req, res) => {
 
 // @desc update contact
 // @route PUT /api/contacts/:id
-// access public
+// access private
 const updateContact = asyncHandler(async (req, res) => {
   const contact = await Contact.findById(req.params.id)
   if (!contact) {
     res.status(404)
     throw new Error("Contact not founc")
   }
+  if (contact.user_id.toString() !== req.user.id) {
+    res.status(403)
+    throw new Error({ message: "user is unauthorize to update" })
+  }
   const updatedContact = await Contact.findByIdAndUpdate(
     req.params.id,
     req.body,
     { new: true }
   )
-  res.status(201).json({ message: `Contact updated for ${req.params.id}` })
+  res.status(201).json(updatedContact)
 })
 
 // @desc delete contact
 // @route DELETE /api/contacts/:id
-// access public
+// access private
 const deleteContact = asyncHandler(async (req, res) => {
   const contact = await Contact.findById(req.params.id)
   if (!contact) {
     res.status(404)
     throw new Error("Contact not founc")
+  }
+  if (contact.user_id.toString() !== req.user.id) {
+    res.status(403)
+    throw new Error({ message: "user is unauthorize to delete" })
   }
   await Contact.deleteOne({ _id: req.params.id })
   res.status(200).json(contact)
